Perform rudimentary "load" and "display" information without UI changes
The ability to easily change the UI is in place though. This is to do
next after getting the future forecast. Then we display in UI cards:
dropdown, current, future weather, dialog that uses stashed data after
loaded in. So it uses data not the immediately returned weather,
storing a current instance to not be forgotten after the enclosed
request.